Move collab code to a separate folder
