+++ NuKeeper.Tests/Engine/CollaborationEngineTests.cs
@@ -8,6 +8,7 @@ using NUnit.Framework;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using NuKeeper.Abstractions.CollaborationModels;
+using NuKeeper.Collaboration;
 
 namespace NuKeeper.Tests.Engine
 {
