+++ NuKeeper/Collaboration/CollaborationEngine.cs
@@ -1,14 +1,15 @@
+using System;
+using System.Threading.Tasks;
 using LibGit2Sharp;
+using NuKeeper.Abstractions.CollaborationModels;
 using NuKeeper.Abstractions.CollaborationPlatform;
 using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Abstractions.Formats;
 using NuKeeper.Abstractions.Logging;
+using NuKeeper.Engine;
 using NuKeeper.Inspection.Files;
-using System;
-using System.Threading.Tasks;
-using NuKeeper.Abstractions.CollaborationModels;
 
-namespace NuKeeper.Engine
+namespace NuKeeper.Collaboration
 {
     public class CollaborationEngine : ICollaborationEngine
     {
