+++ NuKeeper/Commands/CollaborationPlatformNuKeeperCommand.cs
@@ -2,11 +2,11 @@ using McMaster.Extensions.CommandLineUtils;
 using NuKeeper.Abstractions;
 using NuKeeper.Abstractions.CollaborationPlatform;
 using NuKeeper.Abstractions.Configuration;
-using NuKeeper.Engine;
 using NuKeeper.Inspection.Logging;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using NuKeeper.Collaboration;
 
 namespace NuKeeper.Commands
 {
