+++ NuKeeper/Commands/MultipleRepositoryCommand.cs
@@ -4,7 +4,7 @@ using McMaster.Extensions.CommandLineUtils;
 using NuKeeper.Abstractions;
 using NuKeeper.Abstractions.CollaborationPlatform;
 using NuKeeper.Abstractions.Configuration;
-using NuKeeper.Engine;
+using NuKeeper.Collaboration;
 using NuKeeper.Inspection.Logging;
 
 namespace NuKeeper.Commands
