+++ NuKeeper/Commands/OrganisationCommand.cs
@@ -2,7 +2,7 @@ using McMaster.Extensions.CommandLineUtils;
 using NuKeeper.Abstractions;
 using NuKeeper.Abstractions.CollaborationPlatform;
 using NuKeeper.Abstractions.Configuration;
-using NuKeeper.Engine;
+using NuKeeper.Collaboration;
 using NuKeeper.Inspection.Logging;
 
 namespace NuKeeper.Commands
