+++ NuKeeper/Commands/RepositoryCommand.cs
@@ -1,10 +1,10 @@
 using McMaster.Extensions.CommandLineUtils;
 using NuKeeper.Abstractions.CollaborationPlatform;
 using NuKeeper.Abstractions.Configuration;
-using NuKeeper.Engine;
 using NuKeeper.Inspection.Logging;
 using System;
 using System.Collections.Generic;
+using NuKeeper.Collaboration;
 
 namespace NuKeeper.Commands
 {
